pyproject.toml
[project]
name = "langgraph-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.2",
    "langchain-core>=1.1.1",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.1.1",
    "python-dotenv>=1.2.1",
]
